Add files via upload
leetcode 203, Remove Linked List Elements. Done by moving pointers with a dummy node. Struggled a lot with this one, need to revisit it and practice more with Linked Lists. Needed lots of assistance, despite having the right logic. Coding it was difficult; mainly with trying to keep track of the pointers.