Add files via upload
leetcode problem #21. Solved by using a dummy node and comparison loop.
Sidenote: Also noticed that this could be done by merging the problems into a B-Tree, then using an Inorder Traversal to return it sorted in increasing order. Was not able to write that code off the top of my head though, but was a cool nod to my DSA class.